Add multi-window support, refactor pane per workspace, remove PaneGroup and context menus, clean up debug logs
