+++ assets/languages/chinese.json
@@ -10,7 +10,9 @@
   "workspace::CreateOneBlock": "创建新笔记",
   "workspace::ToggleSearchBar": "开关搜索栏",
   "workspace::ToggleSettingsPanel": "开关偏好设置",
+  "workspace::OpenNewWindow": "打开新窗口",
   "new_note": "新笔记",
+  "new_window": "新窗口",
   "search": "搜索",
   "commands": "指令"
 }
