+++ assets/languages/english.json
@@ -10,7 +10,9 @@
   "workspace::CreateOneBlock": "Create a new note",
   "workspace::ToggleSearchBar": "Toggle search bar",
   "workspace::ToggleSettingsPanel": "Toggle settings panel",
+  "workspace::OpenNewWindow": "Open new window",
   "new_note": "New Note",
+  "new_window": "New Window",
   "search": "Search",
   "commands": "Commands"
 }
