+++ crates/opennote-desktop/src/key_mappings/mappings.rs
@@ -14,6 +14,7 @@ actions!(
         ToggleCommandBar,
         ToggleSettingsPanel,
         CreateOneBlock,
+        OpenNewWindow,
         NextTab,
         PreviousTab,
         CloseActiveTab,
@@ -35,6 +36,7 @@ pub fn into_action(context: &str, action: &str) -> Result<Box<dyn Action>> {
             "ToggleCommandBar" => Ok(Box::new(ToggleCommandBar)),
             "ToggleSettingsPanel" => Ok(Box::new(ToggleSettingsPanel)),
             "CreateOneBlock" => Ok(Box::new(CreateOneBlock)),
+            "OpenNewWindow" => Ok(Box::new(OpenNewWindow)),
             "NextTab" => Ok(Box::new(NextTab)),
             "PreviousTab" => Ok(Box::new(PreviousTab)),
             "CloseActiveTab" => Ok(Box::new(CloseActiveTab)),
