+++ crates/opennote-desktop/src/libs/tree.rs
@@ -339,7 +339,6 @@ impl TreeState {
         let block_id = match Uuid::from_str(&self.entries[index].item().id) {
             Ok(result) => result,
             Err(error) => {
-                log::debug!("Failed to open a block with keyboard shortcut: {}", error);
                 return;
             }
         };
@@ -361,7 +360,6 @@ impl TreeState {
         let block_id = match Uuid::from_str(&self.entries[index].item().id) {
             Ok(result) => result,
             Err(error) => {
-                log::debug!("Failed to open a block with keyboard shortcut: {}", error);
                 return;
             }
         };
