+++ crates/opennote-desktop/src/widgets/command_bar/keys_list.rs
@@ -9,7 +9,9 @@ use crate::{
     globals::helpers::get_language_profile,
     key_mappings::{
         helpers::{get_keystrokes_as_shared_string, match_action_to_language},
-        mappings::{CreateOneBlock, ToggleCommandBar, ToggleSettingsPanel, ToggleSidebar},
+        mappings::{
+            CreateOneBlock, OpenNewWindow, ToggleCommandBar, ToggleSettingsPanel, ToggleSidebar,
+        },
     },
 };
 
@@ -28,6 +30,7 @@ impl KeysList {
             Box::new(ToggleCommandBar),
             Box::new(CreateOneBlock),
             Box::new(ToggleSettingsPanel),
+            Box::new(OpenNewWindow),
         ];
 
         let actions_keymaps: Vec<(Box<dyn Action>, Option<SharedString>)> = actions
