+++ crates/opennote-desktop/src/widgets/mod.rs
@@ -1,6 +1,6 @@
 pub mod command_bar;
 pub mod editor;
-pub mod pane;
 pub mod search_bar;
 pub mod sidebar;
 pub mod floating;
+pub mod pane;
