--- crates/opennote-desktop/src/widgets/pane/pane.rs
@@ -1,521 +1,0 @@
-use std::collections::HashMap;
-
-use gpui::{
-    Action, DefiniteLength, Div, DragMoveEvent, Entity, Focusable, Point, SharedString,
-    Subscription, WeakEntity, div,
-};
-use gpui::{Context, FocusHandle, Render, Window, prelude::*};
-use gpui_component::description_list::{DescriptionItem, DescriptionList};
-use gpui_component::{ActiveTheme, Sizable, v_flex};
-use schemars::JsonSchema;
-use serde::Deserialize;
-use uuid::Uuid;
-
-use crate::globals::helpers::get_language_profile;
-use crate::globals::states::States;
-use crate::key_mappings::helpers::get_keystrokes_as_shared_string;
-use crate::key_mappings::mappings::{CreateOneBlock, ToggleCommandBar, ToggleSearchBar};
-use crate::libs::tabs::drag::DraggedItem;
-use crate::libs::tabs::tab_bar::TabBar;
-use crate::widgets::editor::Editor;
-use crate::widgets::pane::pane_group::{PaneGroup, SplitDirection};
-use crate::widgets::pane::tab::{TabState, create_tab_bar_for_blocks};
-use crate::widgets::sidebar::{OpenNoteSidebar, OpenNoteSidebarEvent};
-
-macro_rules! split_structs {
-    ($($name:ident => $doc:literal),* $(,)?) => {
-        $(
-            #[doc = $doc]
-            #[derive(Clone, PartialEq, Debug, Deserialize, Default, Action, JsonSchema)]
-            #[action(namespace = pane)]
-            #[serde(deny_unknown_fields, default)]
-            pub struct $name {
-                pub mode: SplitMode,
-            }
-        )*
-    };
-}
-
-split_structs!(
-    SplitLeft => "Splits the pane to the left.",
-    SplitRight => "Splits the pane to the right.",
-    SplitUp => "Splits the pane upward.",
-    SplitDown => "Splits the pane downward.",
-    SplitHorizontal => "Splits the pane horizontally.",
-    SplitVertical => "Splits the pane vertically."
-);
-
-const DROP_TARGET_SIZE: f32 = 0.2;
-
-#[derive(Clone, Copy, PartialEq, Debug, Deserialize, Default, JsonSchema)]
-#[serde(deny_unknown_fields)]
-pub enum SplitMode {
-    /// Clone the current pane.
-    #[default]
-    ClonePane,
-    /// Create an empty new pane.
-    EmptyPane,
-    /// Move the item into a new pane. This will map to nop if only one pane exists.
-    MovePane,
-}
-
-/// A container for 0 to many items that are open in the workspace.
-/// Treats all items uniformly via the [`ItemHandle`] trait, whether it's an editor, search results multibuffer, terminal or something else,
-/// responsible for managing item tabs, focus and zoom states and drag and drop features.
-/// Can be split, see `PaneGroup` for more details.
-pub struct Pane {
-    pub id: Uuid,
-
-    pub selected_block_id: Option<Uuid>,
-    pub opened_block_ids: Vec<Uuid>,
-    pub opened_block_states: HashMap<Uuid, TabState>,
-    /// The string that will highlighted in the editor
-    pub search_string: Option<SharedString>,
-
-    focus_handle: FocusHandle,
-    pub(crate) editor: Entity<Editor>,
-    drag_split_direction: Option<SplitDirection>,
-    pub(crate) pane_group: WeakEntity<PaneGroup>,
-
-    _subscriptions: Vec<Subscription>,
-}
-
-impl Pane {
-    pub fn new(
-        cx: &mut Context<Self>,
-        window: &mut gpui::Window,
-        pane_group: Entity<PaneGroup>,
-        sidebar: Entity<OpenNoteSidebar>,
-    ) -> Self {
-        let mut _subscriptions = Vec::new();
-
-        let pane_group_clone = pane_group.clone();
-        _subscriptions.push(cx.subscribe(&sidebar, move |this, _entity, event, cx| {
-            if !this.has_opened_blocks() {
-                return;
-            };
-
-            match event {
-                OpenNoteSidebarEvent::BlocksDeleted(block_ids) => {
-                    for id in block_ids {
-                        if this.opened_block_ids.contains(id) {
-                            this.close_tab(id, cx);
-                            let entity = cx.entity();
-
-                            let _ = pane_group_clone.update(cx, |this, cx| {
-                                this.cleanup_pane_without_tabs(entity, cx);
-                            });
-                        }
-                    }
-                }
-            }
-        }));
-
-        let pane_ref = cx.weak_entity();
-
-        Self {
-            id: Uuid::new_v4(),
-            focus_handle: cx.focus_handle(),
-            selected_block_id: None,
-            drag_split_direction: None,
-            search_string: None,
-            editor: cx.new(|cx| Editor::new(cx, window, pane_ref)),
-            opened_block_ids: Vec::new(),
-            opened_block_states: HashMap::new(),
-            pane_group: pane_group.downgrade(),
-            _subscriptions,
-        }
-    }
-
-    pub(crate) fn close_tab(&mut self, block_id: &Uuid, cx: &mut Context<Self>) {
-        // if we have multiple tabs openning
-        if self.opened_block_ids.len() > 1 {
-            // Remove the closed block from the openned blocks,
-            // while also retain an index for moving the focus to the prevoius one
-            let mut removed_index: isize = 0;
-            for (index, opened_block_id) in self.opened_block_ids.iter().enumerate() {
-                if opened_block_id == block_id && index != 0 {
-                    removed_index = index as isize;
-                    break;
-                }
-            }
-
-            self.opened_block_ids.remove(removed_index as usize);
-            self.opened_block_states.remove(block_id);
-
-            // Move the focus to the previous tab / block
-            if let Some(selected_block_id) = &self.selected_block_id {
-                let mut index_to_focus = removed_index - 1;
-
-                // Handle if the closed tab is the first one with no previous tabs
-                if index_to_focus < 0 {
-                    index_to_focus = 0;
-                }
-
-                let Some(block_to_be_selected) = self.opened_block_ids.get(index_to_focus as usize)
-                else {
-                    return;
-                };
-
-                // Move the focus only when the active block has been closed
-                if selected_block_id == block_id {
-                    self.selected_block_id = Some(block_to_be_selected.clone())
-                }
-            }
-
-            cx.notify();
-
-            // Prevent triggering the 1 tab case when
-            // the openned tabs become 1 after the tab closing
-            return;
-        }
-
-        // if we only have 1 tab openning
-        if self.opened_block_ids.len() == 1 {
-            self.opened_block_ids.clear();
-            self.opened_block_states.clear();
-            self.selected_block_id = None;
-
-            cx.notify();
-        }
-
-        // no tab closing for 0 tabs
-    }
-
-    fn handle_drag_move<T: 'static>(
-        &mut self,
-        event: &DragMoveEvent<T>,
-        _window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        let pane_area = event.bounds.size;
-        let cursor_position = event.event.position;
-
-        let relative_cursor = Point::new(
-            event.event.position.x - event.bounds.left(),
-            event.event.position.y - event.bounds.top(),
-        );
-
-        // Relative size of the drop target in the editor that will open dropped file as a split pane (0-0.5)
-        // E.g. 0.25 == If you drop onto the top/bottom quarter of the pane a new vertical split will be used
-        //              If you drop onto the left/right quarter of the pane a new horizontal split will be used
-        //
-        // Referenced from Zed editor's default.json
-        let size = event.bounds.size.width.min(event.bounds.size.height) * DROP_TARGET_SIZE;
-
-        // Reset the drag split direction when the mouse is no longer in this pane
-        if !event.bounds.contains(&cursor_position) {
-            self.drag_split_direction = None;
-            cx.notify();
-            return;
-        }
-
-        let direction = if relative_cursor.x < size
-            || relative_cursor.x > pane_area.width - size
-            || relative_cursor.y < size
-            || relative_cursor.y > pane_area.height - size
-        {
-            [
-                SplitDirection::Up,
-                SplitDirection::Right,
-                SplitDirection::Down,
-                SplitDirection::Left,
-            ]
-            .iter()
-            .min_by_key(|side| match side {
-                SplitDirection::Up => relative_cursor.y,
-                SplitDirection::Right => pane_area.width - relative_cursor.x,
-                SplitDirection::Down => pane_area.height - relative_cursor.y,
-                SplitDirection::Left => relative_cursor.x,
-            })
-            .cloned()
-        } else {
-            None
-        };
-
-        if direction != self.drag_split_direction {
-            self.drag_split_direction = direction;
-        }
-
-        cx.notify();
-    }
-
-    // TODO:
-    // - after splitted, dragging left to right's right should make the pane appears to the right
-    // - active pane should turn to the focused editor's
-    fn handle_item_drop(
-        &mut self,
-        dragged_item: &DraggedItem,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        let Some(owner_pane) = dragged_item.owner_pane.clone() else {
-            return;
-        };
-
-        let Some(owner_pane_id) = dragged_item.owner_pane_id else {
-            return;
-        };
-
-        let Some(dragged_block_id) = dragged_item.block_id else {
-            return;
-        };
-
-        let split_direction = self.drag_split_direction.take();
-        let mut target_pane = cx.entity();
-        let mut old_pane_has_opened_tabs = false;
-
-        // This means the tab is dragged to the same pane but need to split
-        if owner_pane_id == self.id {
-            // Close the dragged tab to create the `move tab` effect
-            self.close_tab(&dragged_block_id, cx);
-            old_pane_has_opened_tabs = self.has_opened_blocks();
-        }
-
-        // Else, it means the tab is dragged to a different pane but need to split
-        if owner_pane_id != self.id {
-            old_pane_has_opened_tabs = owner_pane
-                .update(cx, |this, cx| {
-                    this.close_tab(&dragged_block_id, cx);
-                    this.has_opened_blocks()
-                })
-                .unwrap();
-
-            if let Some(entity) = owner_pane.upgrade() {
-                target_pane = entity;
-            }
-        }
-
-        // Will recursively split
-        // TODO: move the focus to the corresponding pane too
-        let _ = self.pane_group.update(cx, |this, cx| {
-            let pane_group_reference = cx.entity();
-            let new_pane =
-                cx.new(|cx| Pane::new(cx, window, pane_group_reference, this.sidebar.clone()));
-            new_pane.update(cx, |this, cx| {
-                let weak_reference = cx.weak_entity();
-                this.set_selected_block_by_block_id(dragged_block_id, cx);
-                cx.update_global::<States, ()>(|this, _cx| {
-                    this.active_pane = Some(weak_reference);
-                });
-            });
-
-            if let Some(direction) = split_direction {
-                // We had updated the active when creating new pane,
-                // therefore this can be ignored.
-                let _ = this.split(&target_pane, &new_pane, direction, old_pane_has_opened_tabs);
-            }
-
-            cx.notify();
-        });
-
-        cx.notify();
-    }
-
-    pub fn set_search_string(&mut self, string: SharedString) {
-        self.search_string = Some(string)
-    }
-
-    /// `self.search_string` will be emptied, once called
-    pub fn pop_search_string(&mut self) -> Option<SharedString> {
-        self.search_string.take()
-    }
-
-    pub fn set_selected_block_by_block_id(&mut self, block_id: Uuid, cx: &mut Context<Self>) {
-        for opened_block_id in self.opened_block_ids.iter() {
-            if *opened_block_id == block_id {
-                self.selected_block_id = Some(*opened_block_id);
-                cx.notify();
-                return;
-            }
-        }
-
-        self.opened_block_ids.push(block_id);
-        self.opened_block_states.insert(
-            block_id,
-            TabState {
-                ..Default::default()
-            },
-        );
-        self.selected_block_id = Some(block_id);
-        cx.notify();
-    }
-
-    pub fn has_opened_blocks(&self) -> bool {
-        !self.opened_block_ids.is_empty()
-    }
-
-    /// Switch to the next tab (wrapping around).
-    pub fn activate_next_tab(&mut self, cx: &mut Context<Self>) {
-        let current_index = match self.acquire_block_index() {
-            Some(value) => value,
-            None => return,
-        };
-
-        let next_index = if current_index + 1 < self.opened_block_ids.len() {
-            current_index + 1
-        } else {
-            0
-        };
-
-        self.selected_block_id = Some(self.opened_block_ids[next_index]);
-        cx.notify();
-    }
-
-    /// Switch to the previous tab (wrapping around).
-    pub fn activate_previous_tab(&mut self, cx: &mut Context<Self>) {
-        let current_index = match self.acquire_block_index() {
-            Some(value) => value,
-            None => return,
-        };
-
-        let prev_index = if current_index > 0 {
-            current_index - 1
-        } else {
-            self.opened_block_ids.len().saturating_sub(1)
-        };
-
-        self.selected_block_id = Some(self.opened_block_ids[prev_index]);
-        cx.notify();
-    }
-
-    fn acquire_block_index(&mut self) -> Option<usize> {
-        let Some(selected_block_id) = self.selected_block_id else {
-            return None;
-        };
-        let current_index = self
-            .opened_block_ids
-            .iter()
-            .position(|id| *id == selected_block_id);
-        let Some(current_index) = current_index else {
-            return None;
-        };
-        Some(current_index)
-    }
-
-    fn create_commmand_board(cx: &mut Context<'_, Pane>) -> Div {
-        let language_profile = get_language_profile(cx.global(), cx.global()).unwrap();
-
-        v_flex().h_full().child(
-            div().w_48().mx_auto().my_auto().child(
-                DescriptionList::new()
-                    .columns(1)
-                    .bordered(false)
-                    .large()
-                    .children([
-                        DescriptionItem::new(language_profile["search"].to_string()).value(
-                            get_keystrokes_as_shared_string(cx, ToggleSearchBar.boxed_clone())
-                                .unwrap_or("".into()),
-                        ),
-                        DescriptionItem::new(language_profile["commands"].to_string()).value(
-                            get_keystrokes_as_shared_string(cx, ToggleCommandBar.boxed_clone())
-                                .unwrap_or("".into()),
-                        ),
-                        DescriptionItem::new(language_profile["new_note"].to_string()).value(
-                            get_keystrokes_as_shared_string(cx, CreateOneBlock.boxed_clone())
-                                .unwrap_or("".into()),
-                        ),
-                    ]),
-            ),
-        )
-    }
-
-    fn update_editor_with_selected_block(&mut self, cx: &mut Context<'_, Pane>) {
-        if let Some(selected_block_id) = self.selected_block_id {
-            let states: &States = cx.global();
-
-            let block = states.get_block(&selected_block_id);
-
-            if let Some(block) = block {
-                let block = block.to_owned();
-                let search_string = self.pop_search_string();
-
-                self.editor.update(cx, |this, cx| {
-                    // The backend is always the source of truth.
-                    // We fetch the block from the backend with the current uuid.
-                    this.register_block(cx, block);
-                    this.register_highlighted_text(search_string);
-                    cx.notify();
-                });
-            }
-        }
-    }
-}
-
-impl Focusable for Pane {
-    fn focus_handle(&self, _cx: &gpui::App) -> FocusHandle {
-        self.focus_handle.clone()
-    }
-}
-
-impl Render for Pane {
-    fn render(&mut self, _window: &mut Window, cx: &mut Context<Self>) -> impl IntoElement {
-        let base_div = div().flex_1().flex_col(); // We need flex_1 to let the editor to take up the whole space after sidebar disappeared
-
-        // Display search bar, command bar, new doc
-        // and their keyboard shortcuts
-        if self.opened_block_ids.is_empty() {
-            return Self::create_commmand_board(cx);
-        }
-
-        let pane_reference = cx.weak_entity();
-        let pane_id = self.id;
-
-        let tabs: TabBar = create_tab_bar_for_blocks(
-            cx,
-            pane_reference,
-            pane_id,
-            &self.opened_block_ids,
-            self.selected_block_id,
-            &self.opened_block_states,
-        );
-
-        // Open editor only when there is an active block
-        if self.selected_block_id.is_none() {
-            return base_div.child(tabs);
-        };
-
-        self.update_editor_with_selected_block(cx);
-
-        log::debug!("Rendering the pane... {:?}", self.drag_split_direction);
-        base_div.h_full().child(tabs).child(
-            div()
-                .h_full()
-                .on_drag_move::<DraggedItem>(cx.listener(Self::handle_drag_move)) // Calculate the split preview area
-                .child(self.editor.clone())
-                .child(
-                    div()
-                        .absolute()
-                        .bg(cx.theme().blue.opacity(0.3)) // Split preview style
-                        .invisible()
-                        .on_drop(cx.listener(
-                            move |this, dragged_item: &DraggedItem, window, cx| {
-                                this.handle_item_drop(dragged_item, window, cx);
-                            },
-                        ))
-                        .map(|div| {
-                            // Note that we didn't use group drag over like Zed,
-                            // because it didn't work here after several tries.
-                            // Might need to look into this further if there are further issues.
-                            log::debug!("Render split previews...");
-                            let size = DefiniteLength::Fraction(0.5);
-                            match self.drag_split_direction {
-                                None => div.top_0().right_0().bottom_0().left_0(),
-                                Some(SplitDirection::Up) => {
-                                    div.top_0().left_0().right_0().h(size).visible() // Set to visible when dragged over this area
-                                }
-                                Some(SplitDirection::Down) => {
-                                    div.left_0().bottom_0().right_0().h(size).visible()
-                                }
-                                Some(SplitDirection::Left) => {
-                                    div.top_0().left_0().bottom_0().w(size).visible()
-                                }
-                                Some(SplitDirection::Right) => {
-                                    div.top_0().bottom_0().right_0().w(size).visible()
-                                }
-                            }
-                        }),
-                ),
-        )
-    }
-}
