+++ crates/opennote-velotype/src/config/mod.rs
@@ -9,9 +9,8 @@ use serde_json::{Map, Value};
 pub mod preferences;
 
 pub(crate) use preferences::{
-    EditorSettings, ImagePasteBehavior, StartupOpenPreference, apply_configured_language,
-    apply_configured_theme, first_existing_recent_markdown_file, import_language_config_and_select,
-    import_theme_config_and_select, load_or_create_app_preferences, open_preferences_window,
+    EditorSettings, ImagePasteBehavior, apply_configured_language, apply_configured_theme,
+    import_language_config_and_select, import_theme_config_and_select, open_preferences_window,
     read_app_preferences,
 };
 
@@ -60,8 +59,7 @@ impl VelotypeConfigDirs {
     }
 
     pub(crate) fn app_config_file(&self) -> PathBuf {
-        self.root
-            .join("editor_configurations.toml")
+        self.root.join("editor_configurations.toml")
     }
 }
