+++ crates/opennote-velotype/src/editor/persistence.rs
@@ -4,12 +4,9 @@
 //! Markdown. Source mode writes the raw source buffer directly so literal
 //! delimiters are preserved.
 
-use std::path::{Path, PathBuf};
-
 use gpui::*;
 
 use super::Editor;
-use crate::i18n::I18nManager;
 
 fn longest_marker_run(text: &str, marker: char) -> usize {
     let mut longest = 0usize;
@@ -53,173 +50,6 @@ impl Editor {
         } else {
             self.document.markdown_text(cx)
         }
-    }
-
-    pub(super) fn save_dialog_defaults(&self) -> (PathBuf, Option<String>) {
-        if let Some(path) = self.file_path.as_ref() {
-            let directory = path
-                .parent()
-                .map(Path::to_path_buf)
-                .unwrap_or_else(|| std::env::current_dir().unwrap_or_else(|_| PathBuf::from(".")));
-            let suggested_name = path
-                .file_name()
-                .map(|name| name.to_string_lossy().to_string());
-            (directory, suggested_name)
-        } else {
-            (
-                std::env::current_dir().unwrap_or_else(|_| PathBuf::from(".")),
-                Some("untitled.md".to_string()),
-            )
-        }
-    }
-
-    pub(super) fn apply_successful_save(&mut self, path: PathBuf, cx: &mut Context<Self>) {
-        self.file_path = Some(path);
-        self.document_dirty = false;
-        self.pending_window_edited = false;
-        self.pending_window_title_refresh = true;
-        self.pending_close_after_save = false;
-        self.close_dialog_restore_focus = None;
-        self.sync_workspace_after_document_path_change(cx);
-        cx.notify();
-    }
-
-    pub(super) fn save_to_existing_path(
-        &mut self,
-        path: &Path,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) -> bool {
-        let markdown = self.serialized_document_text(cx);
-        match std::fs::write(path, markdown) {
-            Ok(_) => {
-                self.apply_successful_save(path.to_path_buf(), cx);
-                window.set_window_edited(false);
-                true
-            }
-            Err(err) => {
-                let detail = err.to_string();
-                let strings = cx.global::<I18nManager>().strings().clone();
-                let buttons = [strings.info_dialog_ok.as_str()];
-                let _ = window.prompt(
-                    PromptLevel::Critical,
-                    &strings.save_failed_title,
-                    Some(&detail),
-                    &buttons,
-                    cx,
-                );
-                false
-            }
-        }
-    }
-
-    fn save_document_via_prompt(&mut self, window: &mut Window, cx: &mut Context<Self>) {
-        let markdown = self.serialized_document_text(cx);
-        let (default_dir, suggested_name) = self.save_dialog_defaults();
-        let prompt = cx.prompt_for_new_path(&default_dir, suggested_name.as_deref());
-        let weak_editor = cx.entity().downgrade();
-        let weak_editor_for_cancel = weak_editor.clone();
-        let weak_editor_for_error = weak_editor.clone();
-        let weak_editor_for_write_error = weak_editor.clone();
-        let window_handle = window.window_handle();
-        let should_close_after_save = self.pending_close_after_save;
-
-        cx.spawn(async move |_this: WeakEntity<Self>, cx: &mut AsyncApp| {
-            let mut path = match prompt.await {
-                Ok(Ok(Some(path))) => path,
-                Ok(Ok(None)) | Err(_) => {
-                    if should_close_after_save {
-                        let _ = weak_editor_for_cancel
-                            .update(cx, |this, cx| this.abort_pending_close_after_save(cx));
-                    }
-                    return;
-                }
-                Ok(Err(err)) => {
-                    if should_close_after_save {
-                        let _ = weak_editor_for_error
-                            .update(cx, |this, cx| this.abort_pending_close_after_save(cx));
-                    }
-                    let detail = err.to_string();
-                    let _ = cx.update_window(
-                        window_handle,
-                        move |_view: AnyView, window: &mut Window, cx: &mut App| {
-                            let strings = cx.global::<I18nManager>().strings().clone();
-                            let buttons = [strings.info_dialog_ok.as_str()];
-                            let _ = window.prompt(
-                                PromptLevel::Critical,
-                                &strings.save_failed_title,
-                                Some(&detail),
-                                &buttons,
-                                cx,
-                            );
-                        },
-                    );
-                    return;
-                }
-            };
-
-            if path.extension().is_none() {
-                path.set_extension("md");
-            }
-
-            if let Err(err) = std::fs::write(&path, &markdown) {
-                if should_close_after_save {
-                    let _ = weak_editor_for_write_error
-                        .update(cx, |this, cx| this.abort_pending_close_after_save(cx));
-                }
-                let detail = err.to_string();
-                let _ = cx.update_window(
-                    window_handle,
-                    move |_view: AnyView, window: &mut Window, cx: &mut App| {
-                        let strings = cx.global::<I18nManager>().strings().clone();
-                        let buttons = [strings.info_dialog_ok.as_str()];
-                        let _ = window.prompt(
-                            PromptLevel::Critical,
-                            &strings.save_failed_title,
-                            Some(&detail),
-                            &buttons,
-                            cx,
-                        );
-                    },
-                );
-                return;
-            }
-
-            let path_for_state = path.clone();
-            let _ = weak_editor.update(cx, move |this, cx| {
-                this.apply_successful_save(path_for_state, cx);
-            });
-            let _ = cx.update_window(
-                window_handle,
-                move |_view: AnyView, window: &mut Window, _cx: &mut App| {
-                    window.set_window_edited(false);
-                    if should_close_after_save {
-                        window.remove_window();
-                    }
-                },
-            );
-        })
-        .detach();
-    }
-
-    pub(crate) fn save_document(&mut self, window: &mut Window, cx: &mut Context<Self>) {
-        if let Some(path) = self.file_path.clone() {
-            let should_close_after_save = self.pending_close_after_save;
-            if self.save_to_existing_path(&path, window, cx) {
-                if should_close_after_save {
-                    window.remove_window();
-                }
-            } else if should_close_after_save {
-                self.abort_pending_close_after_save(cx);
-            }
-            return;
-        }
-
-        self.save_document_via_prompt(window, cx);
-    }
-
-    pub(crate) fn save_document_as(&mut self, window: &mut Window, cx: &mut Context<Self>) {
-        self.save_document_via_prompt(window, cx);
     }
 }
