+++ crates/opennote-velotype/src/editor/window_state.rs
@@ -153,32 +153,6 @@ impl Editor {
             .min(available_content_width)
     }
 
-    /// Builds the OS window title, including the dirty marker when the
-    /// document has unsaved changes.
-    pub(super) fn window_title(
-        file_path: Option<&Path>,
-        is_dirty: bool,
-        strings: &crate::i18n::I18nStrings,
-    ) -> String {
-        let base_title = if let Some(path) = file_path {
-            format!(
-                "Velotype - {}",
-                path.file_name().map_or_else(
-                    || path.to_string_lossy().to_string(),
-                    |name| name.to_string_lossy().to_string()
-                )
-            )
-        } else {
-            "Velotype".to_string()
-        };
-
-        if is_dirty && !strings.dirty_title_marker.is_empty() {
-            format!("{} {}", strings.dirty_title_marker, base_title)
-        } else {
-            base_title
-        }
-    }
-
     pub(crate) fn on_toggle_view_mode_action(
         &mut self,
         _: &crate::components::ToggleViewMode,
@@ -212,24 +186,6 @@ impl Editor {
         self.redo_document(cx);
     }
 
-    pub(crate) fn on_save_document(
-        &mut self,
-        _: &crate::components::SaveDocument,
-        _window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        self.request_save_document(cx);
-    }
-
-    pub(crate) fn on_save_document_as(
-        &mut self,
-        _: &crate::components::SaveDocumentAs,
-        _window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        self.request_save_document_as(cx);
-    }
-
     pub(crate) fn on_export_html(
         &mut self,
         _: &crate::components::ExportHtml,
@@ -246,24 +202,6 @@ impl Editor {
         cx: &mut Context<Self>,
     ) {
         self.export_document_via_prompt(crate::export::ExportFormat::Pdf, window, cx);
-    }
-
-    pub(crate) fn on_quit_application(
-        &mut self,
-        _: &crate::components::QuitApplication,
-        _window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        crate::app_menu::request_quit_application(cx);
-    }
-
-    pub(crate) fn on_close_window(
-        &mut self,
-        _: &crate::components::CloseWindow,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) {
-        self.request_close_current_window(window, cx);
     }
 
     pub(crate) fn on_install_cli_tool(
@@ -319,7 +257,6 @@ impl Editor {
         self.close_dialog_restore_focus = None;
         self.table_axis_preview = None;
         self.table_axis_selection = None;
-        self.dismiss_contextual_overlays(cx);
         self.sync_table_axis_visuals(cx);
         self.refresh_stable_document_snapshot(cx);
         cx.notify();
@@ -353,10 +290,6 @@ impl Editor {
     }
 
     pub(crate) fn show_info_dialog(&mut self, kind: InfoDialogKind, cx: &mut Context<Self>) {
-        if self.show_unsaved_changes_dialog {
-            return;
-        }
-
         self.menu_bar_open = None;
         self.menu_submenu_open = None;
         self.menu_submenu_panel_hovered = false;
@@ -488,19 +421,19 @@ impl Editor {
         }
     }
 
-    pub(crate) fn request_save_document(&mut self, cx: &mut Context<Self>) {
-        if !self.pending_save {
-            self.pending_save = true;
-            cx.notify();
-        }
-    }
+    // pub(crate) fn request_save_document(&mut self, cx: &mut Context<Self>) {
+    //     if !self.pending_save {
+    //         self.pending_save = true;
+    //         cx.notify();
+    //     }
+    // }
 
-    pub(crate) fn request_save_document_as(&mut self, cx: &mut Context<Self>) {
-        if !self.pending_save_as {
-            self.pending_save_as = true;
-            cx.notify();
-        }
-    }
+    // pub(crate) fn request_save_document_as(&mut self, cx: &mut Context<Self>) {
+    //     if !self.pending_save_as {
+    //         self.pending_save_as = true;
+    //         cx.notify();
+    //     }
+    // }
 
     pub(crate) fn request_open_link_prompt(
         &mut self,
